it's better to make a concern to override the method
the solidus validator class make an override of the initialization and initialize the @field.
