+++ app/models/concerns/solidus_globalize/db_maximum_length_validator.rb
@@ -1,0 +1,19 @@
+module SolidusGlobalize
+  module DbMaximumLengthValidator
+    extend ActiveSupport::Concern
+
+    included do
+
+      def validate(record)
+        field = @field.to_sym
+        limit = record.column_for_attribute(field).limit
+        value = record[field]
+        if value && limit && value.to_s.length > limit
+          record.errors.add(field, :too_long, count: limit)
+        end
+      end
+
+    end
+
+  end
+end
