+++ app/models/spree/validations/db_maximum_length_validator_decorator.rb
@@ -1,14 +1,1 @@
-module Spree
-  module Validations
-    class DbMaximumLengthValidator < ActiveModel::Validator
-      def validate(record)
-        field = @field.to_sym
-        limit = record.column_for_attribute(field).limit
-        value = record[field]
-        if value && limit && value.to_s.length > limit
-          record.errors.add(field, :too_long, count: limit)
-        end
-      end
-    end
-  end
-end
+Spree::Validations::DbMaximumLengthValidator.include(SolidusGlobalize::DbMaximumLengthValidator)
